refactor(cli-plugin-scaffold-graphql-service): split index.ts
