+++ packages/cli-plugin-scaffold-graphql-service/template/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "rimraf ./dist '*.tsbuildinfo' && babel src -d dist --source-maps --copy-files --extensions \".ts,.tsx\" && yarn postbuild",
     "watch": "babel src -d dist --source-maps --copy-files --extensions \".ts,.tsx\" --watch",
-    "postbuild": "cp package.json dist/ && tsc -p tsconfig.build.json"
+    "postbuild": "cp package.json README.md dist/ && tsc -p tsconfig.build.json"
   },
   "dependencies": {
     "@elastic/elasticsearch": "^7.9.1",
