+++ packages/cli-plugin-scaffold-graphql-service/template/src/resolvers/index.ts
@@ -1,0 +1,19 @@
+import getTarget from "./getTarget";
+import listTargets from "./listTargets";
+import isInstalled from "./isInstalled";
+import install from "./install";
+import uninstall from "./uninstall";
+import createTarget from "./createTarget";
+import updateTarget from "./updateTarget";
+import deleteTarget from "./deleteTarget";
+
+export {
+    getTarget,
+    listTargets,
+    isInstalled,
+    install,
+    uninstall,
+    createTarget,
+    updateTarget,
+    deleteTarget
+};
